backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib>=1.7.4
bcrypt==4.1.3
motor==3.3.1
python-jose>=3.3.0
python-multipart>=0.0.9
requests>=2.31.0
